requirements.txt
Flask==1.1.2
Flask-GraphQL==2.0.1
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
Flask-Login==0.5.0
python-keycloak==0.24.0
graphene==2.1.8
graphene-sqlalchemy==2.3.0
Pillow==8.1.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
python-dotenv==0.15.0